% Generated by roxygen2 (4.0.0): do not edit by hand
\name{AIC.admbsecr}
\alias{AIC.admbsecr}
\title{Extract AIC from an admbsecr model object}
\usage{
\method{AIC}{admbsecr}(object, ..., k = 2)
}
\arguments{
\item{object}{A fitted model from \link[admbsecr]{admbsecr}.}

\item{...}{Other parameters (for S3 generic compatibility).}

\item{k}{numeric, the \emph{penalty} per parameter to be used; the
    default \code{k = 2} is the classical AIC.}
}
\description{
Extracts the AIC from an admbsecr model object.
}
\details{
If the model is based on an acoustic survey where there are
multiple calls per individual, then AIC should not be used for
model selection. This function therefore returns NA in this case.
}

