% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-data.r
\name{alldata}
\alias{alldata}
\title{Create an ascr.data object}
\usage{
alldata(
  capt = NULL,
  traps = NULL,
  mask = NULL,
  buffer = NULL,
  cov.var = NULL,
  point.var = NULL,
  cov.list = NULL,
  point.df = NULL,
  nmax = 10,
  maxdist = 1000,
  ...
)
}
\arguments{
\item{capt}{A list containing capture histories and supplementary data for 
each detected individual. It is most easily created using \code{\link{create.capt}}.}

\item{traps}{A matrix with two columns. Each row provides Cartesian 
coordinates for the location of a detector. Alternatively, this can be a 
list of such matrices if detections from multiple detector arrays 
(or 'sessions') are being used to fit a single model.}

\item{mask}{A matrix with two columns, or a list of such matrices for a 
multi-session model. Each row provides Cartesian coordinates for the location
of a mask point. It is most easily created using \code{\link{create.mask}}.
If missing, mask will be created using the provided \code{traps}, 
\code{capt} and \code{buffer}.}

\item{buffer}{The minimum distance between trap locations and the edge of the 
generated mask. This is used with \code{\link{create.mask}} to create the mask
point. Only applied when \code{mask} is missing.}

\item{cov.var}{A vector of character string of the covariate variables used 
to fit the model using \code{\link{fit.ascr}}.}

\item{point.var}{A vector of the character string of the distance covariate
used to fit the model using \code{\link{fit.ascr}}.}

\item{cov.list}{A list of all the covariate data frames. Each row 
provides Cartesian coordinates named "X" and "Y" and its 
covariate values.}

\item{point.df}{A data frames of the distance covariates. Each row 
provides Cartesian coordinates named "X" and "Y" and the 
corresponding feature named "feature".}

\item{nmax}{the number of nearest observations that should be used to predict
the covariate value. This is used with \code{\link{show.prediction}}.
Only applied when \code{cov.var} is numeric variable.}

\item{maxdist}{only observations within a distance of maxdist from the 
prediction location are used to predict the covariate value. This is used 
with \code{\link{show.prediction}}.only applied when \code{cov.var} is numeric
variable.}

\item{...}{Further arguments to be passed to create the mask point using 
\code{\link{create.mask}}.}
}
\value{
An object of class \code{ascr.data}. The object is a list with 
elements \code{capt}, \code{traps}, \code{mask}, \code{cov.prediction.df}, etc.
}
\description{
Sorting the data used for functions in R package \emph{ascr}.
}
\details{
Create a mask object to use with the function \code{\link{fit.ascr}} according 
to the \code{traps} and other argument provided. 

Also, create a list of data frame of the predicted covariate value at each
mask point to use with the function \code{\link{fit.ascr}} when fitting on 
inhomogeneous density surfaces. The function \code{\link{show.prediction}} is
used.
}
