\name{angles}
\alias{angles}
\title{Calculating angles between mask points and traps for SECR models}
\usage{
  angles(traps, mask)
}
\arguments{
  \item{traps}{matrix containing trap coordinates.}

  \item{mask}{matrix containing mask coordinates.}
}
\value{
  A matrix.
}
\description{
  Calculates angles from each trap to each mask point.
}

