% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascr.r
\docType{package}
\name{ascr}
\alias{ascr}
\alias{ascr-package}
\title{ascr: A package to fit acoustic spatial capture-recapture models}
\description{
Spatial capture-recapture (SCR) models estimate animal density from
records of when and were individuals were detected. The first SCR
methods for acoustic data were developed by Efford, Dawson, and
Borchers (2009).
}
\details{
The \code{ascr} package fits SCR models to estimate animal density
from acoustic surveys. In particular, these models can incorporate
additional data like received signal strengths, times of arrival,
estimated distances, and estimated bearings, which are informative
about animals' locations. Other more general software packages
exist for SCR (e.g., the comprehensive \link{secr} R package), but
they cannot handle all data types listed above.

The models that can be fitted in \code{ascr} are described in
Efford, Dawson, and Borchers (2009), Borchers et al (2015), and
Stevenson (2015). The package also implements some unpublished
methods that are described in Stevenson (2016), a PhD thesis.
}
\section{Data structure}{


The easiest way to prepare your data for model fitting is via the
functions \link{create.capt} and \link{create.mask}. Also see the
vignette on data structure, which can be viewed by running
\code{vignette("data-structure")}.
}

\section{Model fitting}{


Model fitting is carried out by the \link{fit.ascr} function.
}

\section{Output and plots}{


Useful information like parameter estimates, standard errors,
confidence intervals, and variance-covariance matrices can be
extracted from a fitted model object using the standard R
functions, like \code{coef}, \code{stdEr}, \code{summary},
\code{confint}, and \code{vcov} (see \link{coef.ascr},
\link{stdEr.ascr}, \link{summary.ascr}, \link{confint.ascr}, and
\link{vcov.ascr}).

Plots of the estimated detection function and of the detection
probability surface can be obtained using \link{show.detfn} and
\link{show.detsurf}, respectively. For inhomogeneous density
models, \link{show.Dsurf} plots the estimated density surface. The
\link{locations} function will plot estimated locations of
individuals or calls.
}

\section{Simulation}{


The \link{sim.capt} function can simulate SCR data, either from
user-specified parameter values or estimated parameters from a
fitted model object.
}

\references{
Borchers, D. L., Stevenson, B. C., Kidney, D., Thomas,
    L., and Marques, T. A. (2015). A unifying model for
    capture-recapture and distance sampling surveys of wildlife
    populations. \emph{Journal of the American Statistical
    Association}, \strong{110}: 195--204.

Efford, M. G., Dawson, D. K., and Borchers,
    D. L. (2009). Population density estimated from locations of
    individuals on a passive detector array. \emph{Ecology},
    \strong{90}: 2676--2682.

Stevenson, B. C., Borchers, D. L., Altwegg, R., Swift,
    R. J., Gillespie, D. M., and Measey, G. J. (2015). A general
    framework for animal density estimation from acoustic
    detections across a fixed microphone array. \emph{Methods in
    Ecology and Evolution}, \strong{6}: 38--48.

Stevenson, B. C. (2016) \emph{Methods in Spatially
    Explicit Capture-Recapture}. PhD thesis, University of St
    Andrews.
}
