\docType{data}
\name{bearing.hn.fit}
\alias{bearing.hn.fit}
\title{An example model object}
\format{A list of class \code{"admbsecr"}.}
\usage{
bearing.hn.fit
}
\description{
This is the model object that results when the
\link[admbsecr]{admbsecr} function is run with
\link[admbsecr]{example.capt}\code{[c("bincapt",
"bearing"]}, \link[admbsecr]{example.traps}, and
\link[admbsecr]{example.mask} set as the arguments
\code{capt}, \code{traps}, and \code{mask}, respectively.
}
\keyword{datasets}

