% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bearings}
\alias{bearings}
\title{Calculating bearings between two sets of points}
\usage{
bearings(a, b)
}
\arguments{
\item{a}{matrix containing a set of coordinates.}

\item{b}{matrix containing another set of coordinates.}
}
\value{
A matrix with bearings of the points in matrix b from the points in matrix a.
}
\description{
Calculates the bearing from one set of points to another.
}
\examples{
bears <- bearings(example.traps, example.mask)
}

