% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot.ascr}
\alias{boot.ascr}
\title{Title}
\usage{
boot.ascr(fit, N = 30, n.cores = 1, infotypes = NULL, seed = NULL)
}
\arguments{
\item{fit}{an object generated by the modelling fit}

\item{N}{a numerical value indicates the number that the bootstrap process will run.}

\item{n.cores}{a numerical value indicates the number of cores to be used. Curretnly this function doesn't support
parallel computing, so it is fixed to be 1.}

\item{infotypes}{a character vector contains the infotypes will be simulated in the additional bootstraps;
or a list with each element as a character vector, used for doing multiple additional
bootstraps with different settings of infotypes.}

\item{seed}{a numerical value indicates the seed for the randomness.}
}
\description{
Title
}
