% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.ascr_tmb}
\alias{confint.ascr_tmb}
\title{Extract confidence interval for ascr.tmb models}
\usage{
\method{confint}{ascr_tmb}(
  object,
  types = NULL,
  level = 0.95,
  pars = NULL,
  new_covariates = NULL,
  ...
)
}
\arguments{
\item{object}{a fitted model from "fit.ascr()".}

\item{types}{a character vector, the same as "coef.ascr_tmb()".}

\item{level}{a numeric value indicates the confident level, default is 0.95.}

\item{pars}{a character vector, the same as "coef.ascr_tmb()".}

\item{new_covariates}{a data frame, the same as "coef.ascr_tmb()".}

\item{...}{}
}
\value{
a matrix
}
\description{
Extract confidence interval for ascr.tmb models
}
