% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert.r
\name{convert.capt}
\alias{convert.capt}
\alias{convert.capt.admbsecr}
\alias{convert.capt.secr}
\title{Capture history conversion.}
\usage{
convert.capt.admbsecr(capt)

convert.capt.secr(capt, traps, capthist = TRUE, cutoff = NULL)
}
\arguments{
\item{capt}{A \code{secr} capture history object for
\code{convert.capt.admbsecr}, or an \code{admbsecr} capture history
object for \code{convert.capt.secr}.}

\item{traps}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a trap (or detector).}

\item{capthist}{Logical, if \code{TRUE}, a \code{capthist} object
is returned. Otherwise a data frame is returned, which is suitable
for the \code{captures} argument to the \link{make.capthist}
function.}

\item{cutoff}{A character string, either \code{"identity"} or
\code{"log"}, which specifies the link function for the signal
strength detection function.}
}
\value{
A capture history object appropriate for analysis using
either the \code{admbsecr} or the \code{secr} package.
}
\description{
These functions convert a capture history object between the
structures required for the \code{admbsecr} and \code{secr}
packages.
}
\examples{
capt <- convert.capt.secr(capt = example$capt, traps = example$traps, cutoff = example$cutoff)
}

