% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert.mask}
\alias{convert.mask}
\title{Convert mask object}
\usage{
convert.mask(mask)
}
\arguments{
\item{mask}{}
}
\value{
An object of class \code{mask}.
}
\description{
Converts an \code{ascr} mask matrix to a \code{secr} mask
object.
}
\details{
The returned object is suitable for use as the \code{mask}
argument of the function \link{secr.fit}.
}
\examples{
mask <- convert.mask(mask = example.data$mask)

}
