% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert.raven}
\alias{convert.raven}
\title{Create a capture history object from a Raven output file}
\usage{
convert.raven(
  dets,
  mics,
  time.range = NULL,
  sound.speed = 330,
  new.allocation = TRUE
)
}
\arguments{
\item{dets}{Detection output dataframe from Raven.}

\item{mics}{A matrix containing the coordinates of microphone
locations.}

\item{time.range}{A vector of length two, providing a range of
times for which a subset should be taken to create the capture
history.}

\item{sound.speed}{The speed of sound in metres per second.}

\item{new.allocation}{Logical, if \code{TRUE}, an improved
call-allocation method is used. The old version is retained so
that older analyses can be replicated.}
}
\description{
Converts a Raven output file to a capture history object
suitable for use with the \link{fit.ascr} function. This uses
\link{make.acoustic.captures} to allocate call identities to
detections.
}
