% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{create.capt}
\alias{create.capt}
\title{Title}
\usage{
create.capt(
  captures,
  traps = NULL,
  ind_model = NULL,
  n.sessions = NULL,
  n.traps = NULL,
  mrds.locs = NULL
)
}
\arguments{
\item{captures}{a list or a data frame with at least 4 columns. The column 1, 2, 4 will be fixed as "session", "ID" and "trap",
each row will be regarded as one detection. Extra information could be provided as columns "bearing", "dist"
,"ss", "toa". And "animal_ID" could be provided if individuals could be distinguished from their acoustic detection.
In addition, any individuals from different sessions will be treated as independent ones.}

\item{traps}{a list or a matrix or a data frame. If it is a data frame or a matrix, it should contains only two columns with the
Cartesian coordinates of the traps. If it is a list, its length will be treated as the number of sessions, and each
element is a data frame or a matrix contains the Cartesian coordinates of the traps in the corresponding session.
If it is a list with only one element, or a data frame, or a matrix, the number of session will be determined
by the argument "n.sessions" or max(captures$session) if "n.sessions" is not provided. This setting is prepared 
for the scenario that the survey is conducted with the same detectors for multiple sessions.}

\item{ind_model}{a logical value indicates whether to include individual identification which should be recorded as "animal_ID",
a column in the argument "captures". By default, it will be NULL, so the model will determine it automatically
depend on whether this information is provided in "captures".}

\item{n.sessions}{a numeric value denotes the number of sessions. It will only be used when the argument "traps" is not provided,
or it is a list with one element, or it is a data frame, or it is a matrix.}

\item{n.traps}{a numeric vector denotes the number of traps in each session. It will only be used when the argument "traps" is not provided.
If its length is 1, its value will be used for each session when we have multiple sessions. If its length is greater than 1,
the length of it must match the number of sessions.}

\item{mrds.locs}{a list with length of n.sessions with data frames or matrices as components, each data.frame or matrix
contains two columns record the Cartesian coordinates of each call. If a session has no detection,
keep the corresponding component as NULL. If animal.model, then each component must be a data.frame
with 4 columns: "animal_ID", "ID", "mrds_x" and "mrds_y".}
}
\value{
an object ready for model fitting.
}
\description{
Title
}
