% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{demo_fit}
\alias{demo_fit}
\title{Show demo}
\usage{
demo_fit(data_name, fit = TRUE, gradient_free = FALSE, sv_link = NULL)
}
\arguments{
\item{data_name}{}

\item{fit}{}

\item{gradient_free}{}

\item{sv_link}{}
}
\description{
Show demo
}
