\docType{data}
\name{example.capt}
\alias{example.capt}
\title{An example capture history object}
\format{A list, which is the correct format for use as the
\code{capt} argument to the function \link[admbsecr]{admbsecr}.}
\usage{
example.capt
}
\description{
A list containing various additional information types.
These data were simulated using \link[admbsecr]{sim.capt}
using the trap locations in \link[admbsecr]{example.traps}.
}
\keyword{datasets}

