% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{example.mask}
\alias{example.mask}
\title{An example mask object}
\format{A matrix with two columns. Each row gives the Cartesian
coordinates of a mask point.}
\usage{
example.mask
}
\description{
A matrix containing mask point locations. These mask point
locations are suitable for analysis of the data \link{example.capt}
using the function \link{admbsecr}.
}
\keyword{datasets}

