% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{example}
\alias{example}
\title{Example data}
\format{A list.}
\usage{
example
}
\description{
This object contains simulated data with all types of supplementary
information, corresponding trap locations, and a suitable mask
object. Also included are some example model fits, which were
generated from these data using the \link{admbsecr} function.
}
\details{
This object is a list which contains components:
\itemize{
\item \code{capt}: A capture history object.
\item \code{traps}: A traps object.
\item \code{mask}: A suitable mask object.
\item \code{cutoff}: The cutoff value used to simluate these data.
\item \code{fits}: Some example model fits.
}
}
\keyword{datasets}

