\docType{data}
\name{example.traps}
\alias{example.traps}
\title{An example traps object}
\format{A matrix with two columns. Each row gives the Cartesian
coordinates of a trap.}
\usage{
example.traps
}
\description{
A matrix containing the trap locations used for the
simulation of the data \link{example.capt}. This object is
suitable for use as the \code{traps} argument of the
function \link{admbsecr}.
}
\keyword{datasets}

