% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{expand.polygons}
\alias{expand.polygons}
\title{Function to add covariates for each polygon created by the polygonize function.}
\usage{
expand.polygons(from, mask, raster)
}
\arguments{
\item{from}{Dataframe of trap locations.}

\item{mask}{A mask object created with \code{\link{create.mask}}.}

\item{raster}{A list of raster object to calculate covariate statistics for each grid cell.}
}
\value{
Buffered grid with covariate values for each grid cell (i.e. SpatialPolygonsDataFrame).
}
\description{
The function extracts mean covariate values for each grid cell of the buffered spatial grid. In order to be efficient the function creates the buffered grid and a clone of the unbuffered grid. It pastes covariates of the mask in the unbuffered grid, it calculates the difference between unbuffered grid and buffered grid, and calcultes covariate values only for that difference. It then merges the unbuffered grid with the difference to get final buffered grid with associated covariate values. Buffer is set to 3x the grid resolution.
}
