% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.bias}
\alias{get.bias}
\title{Extracting estimated biases}
\usage{
get.bias(fit, pars = "fitted", mce = FALSE)
}
\description{
Extracts bias in parameter estimates, calculated using the
bootstrap procedure carried out by \link{boot.ascr}.
}
\seealso{
\link{boot.ascr} for the bootstrap procedure.

\link{get.mce} for Monte Carlo error the biases are
    subject to.
}
