% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.capt.hist}
\alias{get.capt.hist}
\title{Function to turn data frame into capture history matrix}
\usage{
get.capt.hist(data)
}
\arguments{
\item{data}{A data.frame with at least the following named columns:
\code{occasion}, numeric day or time index;
\code{post}, a factor vector of post (trap) names,
and \code{group}. It is the pairing of
\code{group} and \code{occasion} that forms the individual
"call id". Optional to include a column of bearings (in radians)
and/or distances.}
}
\description{
Function to turn data frame into capture history matrix
}
