% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.mce}
\alias{get.mce}
\title{Extracting Monte Carlo error}
\usage{
get.mce(fit, estimate)
}
\arguments{
\item{fit}{A fitted model from \link{admbsecr}.}

\item{estimate}{A character string, either \code{"bias"} or
\code{"se"}, which determines whether Monte Carlo errors for
bias estimates or standard errors are reported.}
}
\description{
Extracts calculated Monte Carlo errors from a bootstrap procedure
carried out by \link{boot.admbsecr}.
}
\seealso{
\link{boot.admbsecr} for the bootstrap procedure.

\link{stdEr.admbsecr.boot} for standard errors.

\link{get.bias} for estimated biases.
}

