% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.r
\name{launch.ui}
\alias{launch.ui}
\title{Function to launch ascr shniy application.
Looks for the ascr app within the package directory.
No arguments are required.}
\usage{
launch.ui()
}
\description{
Function to launch ascr shniy application.
Looks for the ascr app within the package directory.
No arguments are required.
}
