\docType{data}
\name{lightfooti.capt}
\alias{lightfooti.capt}
\title{\emph{Arthroleptella lightfooti} capture history object}
\format{A list, which is the correct format for use as the
\code{capt} argument to the function \link{admbsecr}.}
\usage{
lightfooti.capt
}
\description{
A capture history object from an acoustic survey of the
Western Cape moss frog \emph{Arthroleptella lightfooti}.
These data are from a 25 s subset of the original
recording, taken on 16 May 2012 at Silvermine, Table
Mountain National Park, Cape Town, South Africa.
}
\details{
Acoustic signal strengths and times of arrival were
measured, and this information is contained in this capture
history object.
}
\keyword{datasets}

