\docType{data}
\name{lightfooti}
\alias{lightfooti}
\title{\emph{Arthroleptella lightfooti} survey data}
\format{A list.}
\usage{
lightfooti
}
\description{
Data from an acoustic survey of the Western Cape moss frog
\emph{Arthroleptella lightfooti}. These data are from a 25
s subset of the original recording, taken on 16 May 2012 at
Silvermine, Table Mountain National Park, Cape Town, South
Africa. Acoustic signal strengths and times of arrival were
measured, and this information is contained in the capture
history object.
}
\details{
This object is a list which contains components: \itemize{
\item \code{capt}: A capture history object. \item
\code{traps}: A traps object. \item \code{mask}: A suitable
mask object. \item \code{cutoff}: The microphone cutoff
value. \item \code{freqs}: A vector of call frequencies
measured independently to the acoustic survey. }
}
\keyword{datasets}

