% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{make.dm}
\alias{make.dm}
\title{Function to create the design matrix for GAM models.}
\usage{
make.dm(model, data, knots)
}
\arguments{
\item{model}{A GAM model to be specified as a combination of covariates.}

\item{data}{Dataframe of covariate values. Note that each column of the Dataframe should be a covariate and name should correspond to the name spacified in \code{model}.}

\item{knots}{A list of knot positions for each covariate. For more information visit \code{\link[mgcv]{smooth.construct}}.}
}
\value{
A Dataframe of smooth terms for each covariate.
}
\description{
The function uses information about the model structure to create a smooth object through the mgcv Package.
}
