% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mask.test}
\alias{mask.test}
\title{Testing the mask for first call models under a set of parameter values.}
\usage{
mask.test(mask, traps, pars, cutoff, lower.cutoff)

mask.test(mask, traps, pars, cutoff, lower.cutoff)
}
\description{
Testing the mask for first call models under a set of parameter values.

Testing the mask for first call models under a set of parameter values.
}

