% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascr.r
\docType{data}
\name{multi.example.data}
\alias{multi.example.data}
\title{Multi-array example data}
\format{
A list.
}
\usage{
multi.example.data
}
\description{
This object contains simulated data From multiple detector
arrays.
}
\details{
This object is a list which contains components:
\itemize{

\item \code{capt}: A list of capture history objects, one from each
detector array.

\item \code{traps}: A list of traps objects.
\item \code{mask}: A list of suitable mask objects.
}
}
\keyword{datasets}
