% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{myDist}
\alias{myDist}
\title{Function to calculate cost and commute distances.}
\usage{
myDist(par, exp.poly = NULL, from, mask, trans.fn, raster, model,
  knots = NULL, comm.dist = FALSE, parallel = FALSE, ncores = NULL)
}
\arguments{
\item{par}{Non-euclidean parameters that are automatically generated by the optimization algorithm \code{\link[stats]{optim}}.}

\item{exp.poly}{SpatialPolygonDataFrame generated by \code{\link{expand.polygons}}.}

\item{from}{Dataframe of trap locations.}

\item{mask}{A mask object created with \code{\link{create.mask}}.}

\item{trans.fn}{The transition function for the calculation of cost and commute distances. For more information see \code{\link[gdistance]{transition}}.}

\item{raster}{List of raster objects for each covariate.}

\item{model}{A GAM model to be specified as a combination of covariates.}

\item{knots}{A list of knot positions for each covariate. For more information visit \code{\link[mgcv]{smooth.construct}}.}

\item{comm.dist}{TRUE for the calculation of commute distances. FALSE for the calculation of cost distances. Default is FALSE.}

\item{parallel}{TRUE for the parallelization of commute distances calculation. Default is FALSE.}

\item{ncores}{Number of cores if \code{parallel=TRUE}.}
}
\value{
An array for distances from each trap to each grid cell.
}
\description{
This function calculates cost and commute distances that are used to fit ascr models. Commute distances can be calculated on multiple cores to speed up the process.
}
