% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{p.dot}
\alias{p.dot}
\title{Calculating detection probabilities.}
\usage{
p.dot(fit = NULL, session = 1, esa = FALSE, points = get.mask(fit,
  session), traps = NULL, detfn = NULL, ss.link = NULL,
  pars = NULL, n.quadpoints = 8)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{session}{For multisession models, the session from which the
trap locations should be taken.}

\item{esa}{Logical, if \code{TRUE} the effective sampling area is
returned instead of capture probabilities.}

\item{points}{A matrix with two columns. Each row provides
Cartesian coordinates for the location of a point at which a
capture probability should be returned.}

\item{traps}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a trap (or detector). Ignored
if \code{fit} is not \code{NULL}.}

\item{detfn}{A character string specifying the detection function
to be used. One of "hn" (halfnormal), "hr" (hazard rate), "th"
(threshold), "lth" (log-link threshold), or "ss" (signal
strength). Ignored if \code{fit} is not \code{NULL}.}

\item{ss.link}{A character string, either \code{"identity"},
\code{"log"}, or \code{"spherical"}, which specifies the
relationship between the expected received signal strength and
distance from the microphone. See the documentation for
\link{fit.ascr} for further details. Ignored if \code{fit} is
not \code{NULL}.}

\item{pars}{A named list. Component names are parameter names, and
each component is a value for the associated parameter. Ignored
if \code{fit} is not \code{NULL}.}

\item{n.quadpoints}{An integer, giving the number of quadrature
points used for numerical integration over the possible call
directions.}
}
\value{
A vector containing detection probabilities for each
    location in \code{points}.
}
\description{
Calculates the probability of detection by at least one detector
for specific locations in the survey area.
}
