% Generated by roxygen2 (4.0.2): do not edit by hand
\name{par.admbsecr}
\alias{par.admbsecr}
\title{Parallelising admbsecr fits using a cluster}
\usage{
par.admbsecr(n.cores, ...)
}
\arguments{
\item{n.cores}{A positive integer representing the number of cores
to use for parallel processing.}

\item{...}{Lists with components comprising arguments for a call to
\link{admbsecr}. Component names must be the argument names.}
}
\value{
A list, where components are objects returned by
\link{admbsecr}. There is one component for each list of arguments
provide in \code{...}.
}
\description{
Fits SECR models on different cores within a cluster.
}
\examples{
\dontrun{
## Running the examples in the admbsecr() documentation in parallel.
simple.capt <- example$capt["bincapt"]
simple.hn.args <- list(capt = simple.capt, traps = example$traps,
                       mask = example$mask, fix = list(g0 = 1))
simple.hr.args <- list(capt = simple.capt, traps = example$traps,
                       mask = example$mask, detfn = "hr")
bearing.capt <- example.capt[c("bincapt", "bearing")]
bearing.hn.args <- list(capt = bearing.capt, traps = example$traps,
                        mask = example$mask, fix = list(g0 = 1))
## This will only run if you have 4 cores available, you may need
## to alter n.cores as appropriate.
fits <- par.admbsecr(n.cores = 4, simple.hn.args, simple.hr.args,
                     bearing.hn.args)
}
}

