% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_in_fit.R
\name{par.extend.fun}
\alias{par.extend.fun}
\title{fgam' is an object for compatibility with ADMB model
ID',}
\usage{
par.extend.fun(
  par.extend = par.extend,
  data.full = data.full,
  data.mask = data.mask,
  dims = dims,
  fulllist.par = fulllist.par
)
}
\description{
fgam' is an object for compatibility with ADMB model
ID',
}
