% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{par.fit.ascr}
\alias{par.fit.ascr}
\alias{par.admbsecr}
\title{Parallelising ascr fits using a cluster}
\usage{
par.fit.ascr(n.cores, ..., arg.list = NULL)

par.admbsecr(n.cores, ..., arg.list = NULL)
}
\arguments{
\item{n.cores}{A positive integer representing the number of cores
to use for parallel processing.}

\item{...}{Lists with components comprising arguments for a call to
\link{fit.ascr}. Component names must be the argument names.}

\item{arg.list}{Alternatively, a list with components comprising
the lists of arguments, as above.}
}
\value{
A list, where components are objects returned by
\link{fit.ascr}. There is one component for each list of arguments
provide in \code{...}.
}
\description{
Fits SECR models on different cores within a cluster.
}
\examples{
\dontrun{
## Running the examples in the fit.ascr() documentation in parallel.
simple.capt <- example$capt["bincapt"]
simple.hn.args <- list(capt = simple.capt, traps = example$traps,
                       mask = example$mask, fix = list(g0 = 1))
simple.hr.args <- list(capt = simple.capt, traps = example$traps,
                       mask = example$mask, detfn = "hr")
bearing.capt <- example$capt[c("bincapt", "bearing")]
bearing.hn.args <- list(capt = bearing.capt, traps = example$traps,
                        mask = example$mask, fix = list(g0 = 1))
## This will only run if you have 4 cores available, you may need
## to alter n.cores as appropriate.
fits <- par.fit.ascr(n.cores = 4, simple.hn.args, simple.hr.args,
                     bearing.hn.args)
}

}
