% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ascr-data.r
\name{plot.ascrdata}
\alias{plot.ascrdata}
\title{Plot an ascrdata object}
\usage{
\method{plot}{ascrdata}(datalist, type = "covariate", session = 1, cov = "all")
}
\arguments{
\item{datalist}{A list containing capture histories, traps locations, etc.
It is most easily created using \link{alldata}.}

\item{type}{A character string specifying the type of plot to be generated.
One of "capt", "survey" or "covariate".}

\item{session}{The session(s) for which the mask point and trap
locations are to be plotted.}

\item{cov}{A character string specifying the covariate to be plotted. 
Only used when \code{covariate} is chosen for \code{type}.
By default, "all" will plot all covariates stored in \code{datalist}.}
}
\description{
\description{
Generate several types of plots from an ascrdata object including: 
  
  1. trap locations and capture histories using \link{show.capt}.
  
  2. mask and trap layout using \link{show.survey}.
        
  3. the predicted covariate values at each mask point using 
     \link{show.prediction}. See 'Details' for further information.
}
}
\details{
When \code{covariate} is chosen for \code{type}, \code{datalist}
should at least contain either \code{cov.var} or \code{point.var} and the
corresponding \code{cov.list} or \code{point.df}. See \link{alldata} for details.
}
