% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{plot.distgam}
\alias{plot.distgam}
\title{Function to plotGamma distridution of distance measurtment error
from model.}
\usage{
\method{plot}{distgam}(fit = NULL, d = NULL)
}
\arguments{
\item{fit}{ascr model}

\item{d}{distance of call/animal at which to plot}
}
\description{
Function to plotGamma distridution of distance measurtment error
from model.
}
