% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{plot.dvm}
\alias{plot.dvm}
\title{Function to plot Von Mises distribution of bearing measurement error
from model.}
\usage{
\method{plot}{dvm}(fit = NULL)
}
\arguments{
\item{fit}{ascr model}
}
\description{
Function to plot Von Mises distribution of bearing measurement error
from model.
}
