% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{polygonize}
\alias{polygonize}
\title{Function to buffer and polygonize the mask.}
\usage{
polygonize(traps, mask, buffer, resolution, raster)
}
\arguments{
\item{traps}{Dataframe of trap locations.}

\item{mask}{A mask object created with \code{\link{create.mask}}.}

\item{buffer}{Distance in meters to buffer the mask.}

\item{resolution}{Resolution in meters of the mask cells.}

\item{raster}{A raster object with information about CRS and extent.}
}
\value{
A buffered mask as SpatialPolygons.
}
\description{
This function is compulsory for the calculation of commute distances. It creates a buffered grid as SpatialPolygons.
}
