% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.coef_ascr_tmb}
\alias{print.coef_ascr_tmb}
\title{S3 method for output of coef.ascr_tmb and coef.ascr_boot}
\usage{
\method{print}{coef_ascr_tmb}(x)
}
\arguments{
\item{x}{an object generated by the function "coef.ascr_tmb()" or "coef.ascr_boot()".}
}
\description{
S3 method for output of coef.ascr_tmb and coef.ascr_boot
}
