\name{print.summary.admbsecr}
\alias{print.summary.admbsecr}
\title{Printing admbsecr summaries}
\usage{
\method{print}{summary.admbsecr}(x, ...)
}
\arguments{
  \item{x}{An object of class \code{summary.admbsecr}.}

  \item{...}{Other parameters (for S3 generic
  compatibility).}
}
\description{
Printing admbsecr summaries
}

