% Generated by roxygen2 (4.0.2): do not edit by hand
\name{secr_nll}
\alias{secr_nll}
\title{Evaluating the likelihood using C++}
\usage{
secr_nll(link_pars, dat)
}
\arguments{
\item{link_pars}{Parameter values on link scale.}

\item{dat}{Data list.}
}
\value{
The value of the negative log-likelihood.
}
\description{
Returns the likelihood for a vector of parameter values.
}

