% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.r
\name{show.capt}
\alias{show.capt}
\title{Plotting capture histories}
\usage{
show.capt(capt, traps, mask, session = NULL, ask = TRUE)
}
\arguments{
\item{capt}{A list containing capture histories and supplementary
data for each detected individual. It is most easily created
using \link{create.capt}.}

\item{traps}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a detector. Alternatively, this
can be a list of such matrices if detections from multiple
detector arrays (or `sessions') are being used to fit a single
model.}

\item{mask}{A matrix with two columns, or a list of such matrices
for a multi-session model. Each row provides Cartesian
coordinates for the location of a mask point. It is most easily
created using \link{create.mask}.}

\item{session}{The session for which to plot capture histories. If
\code{NULL}, the default, then plots will be shown for all
sessions.}

\item{ask}{If \code{TRUE}, the user will be prompted to press
return to see the next plot.}
}
\description{
Plots trap locations and capture histories.
}
