% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.r
\name{show.Dsurf}
\alias{show.Dsurf}
\title{Plotting an estimated density surface.}
\usage{
show.Dsurf(fit, session = 1)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{session}{The session with the detector array and invidual(s)
to be plotted (for multi-session models only).}
}
\description{
Plots density surface estimated by a model fitted with the function
\link{fit.ascr}.
}
\examples{
simple.capt <- example$capt[1]
cov.df <- data.frame(x = example$mask[, 1]/1000, y = example$mask[, 2]/1000)
fit <- fit.ascr(capt = simple.capt, traps = example$traps, mask = example$mask,
                fix = list(g0 = 1), ihd.opts = list(model = ~ x + y,
                                                    covariates = cov.df))
show.Dsurf(fit)
}
