% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{show.mask}
\alias{show.mask}
\title{Function to plot mask along with trap locations in a 'tidy' presentable manner}
\usage{
show.mask(mask = NULL, traps = NULL)
}
\arguments{
\item{mask}{mask object from \code{create.mask}}

\item{traps}{a matrix of trap locations used to create the \link{mask}}
}
\description{
Function to plot mask along with trap locations in a 'tidy' presentable manner
}
