% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-function.R
\name{show.prediction}
\alias{show.prediction}
\title{Interpolation for covariates}
\usage{
show.prediction(
  mask,
  cov.var = NULL,
  point.var = NULL,
  nmax = 10,
  maxdist = 1000,
  cov.list = NULL,
  point.df = NULL
)
}
\arguments{
\item{mask}{A matrix with two columns. Each row provides Cartesian 
coordinates named "x" and "y" for the location of a mask point.}

\item{cov.var}{A vector of the covariate variables for prediction.}

\item{point.var}{A vector of the distance covariate for prediction.}

\item{nmax}{the number of nearest observations that should be used for
prediction. This is only applied to numeric covariate variable.}

\item{maxdist}{only observations within a distance of maxdist from the 
prediction location are used for prediction; This is only applied to numeric 
covariate variable.}

\item{cov.list}{A list of all the covariate data frames. Each row 
provides Cartesian coordinates named "X" and "Y" and its 
covariate values.}

\item{point.df}{A data frames of the distance covariates. Each row 
provides Cartesian coordinates named "X" and "Y" and the 
corresponding feature named "feature".}
}
\value{
A list with elements \code{prediction} and \code{plot} is returned.
}
\description{
Interpolates the covariate for the mask points using 
inverse distance weighted (IDW) method.
}
\details{
IDW uses the measured values surrounding the prediction location to predict
a value for any unmeasured location. 

Each measured point, according to IDW, has a local influence that decreases
with distance. It gives more weight to points that are closest to the
prediction location and the weights decrease as the distance increases.
}
