% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{simple.hn.fit}
\alias{simple.hn.fit}
\title{An example model object}
\format{A list of class \code{"admbsecr"}.}
\usage{
simple.hn.fit
}
\description{
This is the model object that results when the \link{admbsecr}
function is run with \link{example.capt}\code{["bincapt"]},
\link{example.traps}, and \link{example.mask} set as the
arguments \code{capt}, \code{traps}, and \code{mask}, respectively.
}
\keyword{datasets}

