\docType{data}
\name{simple.hr.fit}
\alias{simple.hr.fit}
\title{An example model object}
\format{A list of class \code{"admbsecr"}.}
\usage{
simple.hr.fit
}
\description{
This is the model object that results when the
\link[admbsecr]{admbsecr} function is run with
\link[admbsecr]{example.capt}\code{["bincapt"]},
\link[admbsecr]{example.traps},
\link[admbsecr]{example.mask}, and \code{"hr"} set as the
arguments \code{capt}, \code{traps}, \code{mask}, and
\code{detfn}, respectively.
}
\keyword{datasets}

