% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{stdEr.ascr.boot}
\alias{stdEr.ascr.boot}
\title{Extract standard errors from a bootstrapped ascr model object}
\usage{
\method{stdEr}{ascr.boot}(object, pars = "fitted", mce = FALSE, ...)
}
\arguments{
\item{object}{A fitted model from \link[ascr]{fit.ascr}.}

\item{pars}{A character vector containing either parameter names,
or a subset of \code{"all"}, \code{"derived"}, \code{"fitted"},
and \code{"linked"}; \code{"fitted"} corresponds to the
parameters of interest, \code{"derived"} corresponds to
quantities that are functions of these parameters (e.g., the
effective survey area or animal density from an acoustic
survey), and \code{"linked"} corresponds to the parameters AD
Model Builder has maximised the likelihood over.}

\item{mce}{Logical, if \code{TRUE} Monte Carlo error for the
standard errors is also returned.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Extracts standard errors for parameters of a model fitted using
\link[ascr]{fit.ascr}, with a bootstrap procedure carried out
using \link[ascr]{boot.ascr}.
}
