% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{stdEr.ascr}
\alias{stdEr.ascr}
\title{Extract standard errors from an ascr model fit}
\usage{
\method{stdEr}{ascr}(object, pars = "fitted", ...)
}
\arguments{
\item{object}{A fitted model from \link[ascr]{fit.ascr}.}

\item{pars}{A character vector containing either parameter names,
or a subset of \code{"all"}, \code{"derived"}, \code{"fitted"},
and \code{"linked"}; \code{"fitted"} corresponds to the
parameters of interest, \code{"derived"} corresponds to
quantities that are functions of these parameters (e.g., the
effective survey area or animal density from an acoustic
survey), and \code{"linked"} corresponds to the parameters AD
Model Builder has maximised the likelihood over.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Extracts standard errors for estimated and derived parameters from
a model fitted using \link[ascr]{fit.ascr}.
}
\examples{
stdEr(example$fits$simple.hn)
stdEr(example$fits$simple.hn, pars = "all")
stdEr(example$fits$simple.hn, pars = "derived")

}
