% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stdEr.ascr_tmb}
\alias{stdEr.ascr_tmb}
\title{Extract standard errors of the estimated parameters from ascr models}
\usage{
\method{stdEr}{ascr_tmb}(
  object,
  types = NULL,
  pars = NULL,
  new_covariates = NULL,
  show_fixed_par = TRUE,
  ...
)
}
\arguments{
\item{object}{a fitted model from "fit.ascr()".}

\item{types}{a character vector, the same as "coef.ascr_tmb()".}

\item{pars}{a character vector, the same as "coef.ascr_tmb()".}

\item{new_covariates}{a data frame, the same as "coef.ascr_tmb()".}

\item{show_fixed_par}{a logical value, the same as "vcov.ascr_tmb()".}

\item{...}{}
}
\value{
a named numeric vector
}
\description{
Extract standard errors of the estimated parameters from ascr models
}
