% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.r
\name{subsample.se}
\alias{subsample.se}
\title{Combining subsamples to obtain a standard error.}
\usage{
subsample.se(..., par, plot = TRUE, ceiling = NULL)
}
\arguments{
\item{...}{A number of bootstrap model objects.}

\item{par}{A character string providing the parameter for which to
calculate a standard error.}

\item{plot}{Logical, if \code{TRUE}, a boxplot is produced.}

\item{ceiling}{A threshold value; bootstrapped parameter values
above this are discarded.}
}
\description{
Calculates a single standard error for a parameter that has been
calculated by averaging over subsamples.
}
