% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{summary.ascr}
\alias{summary.ascr}
\title{Summarising ascr model fits}
\usage{
\method{summary}{ascr}(object, ...)
}
\arguments{
\item{object}{A fitted model from \link[ascr]{fit.ascr}.}

\item{...}{Other parameters (for S3 generic compatibility).}
}
\description{
Provides a useful summary of the model fit.
}
