% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test.admbsecr}
\alias{test.admbsecr}
\title{Run admbsecr tests}
\usage{
test.admbsecr()
}
\description{
Runs tests for admbsecr.
}
\details{
This function was written to allow users to test their admbsecr
installation. Users cannot use the \link[testthat]{test_package}
function from the testthat package as the tests are not
installed. This is because tests fail on the R-forge servers due to
the AD model builder executable.
}

