% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{test.ascr}
\alias{test.ascr}
\alias{test.admbsecr}
\title{Run ascr tests}
\usage{
test.ascr(quick = FALSE)

test.admbsecr(quick = FALSE)
}
\arguments{
\item{quick}{Logical, if \code{TRUE}, only a quick check is carried
out that tests whether or not the AD Model Builder executable
is running correctly.}
}
\description{
Runs tests for ascr.
}
\details{
This function was written to allow users to test their ascr
installation. Users cannot use the \link[testthat]{test_package}
function from the testthat package as the tests are not
installed. This is because tests fail on the R-forge servers due to
the AD model builder executable.
}
