\name{toa.ssq}
\alias{toa.ssq}
\title{Sum of squares TOA matrix}
\usage{
  toa.ssq(wit, dists, speed)
}
\arguments{
  \item{wit}{capture history.}

  \item{dists}{distances.}

  \item{speed}{speed of sound.}
}
\value{
  A matrix.
}
\description{
  Calculates ssqtoa matrix for a SECR model with TOA
  information.
}

