% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tmb.R
\name{vcov.ascr_tmb}
\alias{vcov.ascr_tmb}
\title{og' below means original output from the object}
\usage{
\method{vcov}{ascr_tmb}(object, types = NULL, pars = NULL, new_covariates = NULL, ...)
}
\description{
og' below means original output from the object
}
