% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{bearing.hn.fit}
\alias{bearing.hn.fit}
\title{An example model object}
\format{A list of class \code{"admbsecr"}.}
\usage{
bearing.hn.fit
}
\description{
This is the model object that results when the
\link{admbsecr} function is run with
\link{example.capt}\code{[c("bincapt", "bearing"]},
\link{example.traps}, and \link{example.mask}
set as the arguments \code{capt}, \code{traps}, and \code{mask},
respectively.
}
\keyword{datasets}

