% Generated by roxygen2 (4.0.0): do not edit by hand
\name{boot.admbsecr}
\alias{boot.admbsecr}
\title{Bootstrapping SECR data}
\usage{
boot.admbsecr(fit, N, prog = TRUE, n.cores = 1, M = 10000,
  infotypes = NULL)
}
\arguments{
\item{fit}{A fitted \code{admbsecr} model object.}

\item{N}{The number of bootstrap resamples.}

\item{prog}{Logical, if \code{TRUE}, a progress bar is shown. Only
available if \code{n.cores} is 1.}

\item{n.cores}{A positive integer representing the number of cores
to use for parallel processing.}

\item{M}{The number of bootstrap resamples for the secondary
bootstrap used to calculate Monte Carlo error. See 'Details'
below. If M = 0, then this is skipped.}

\item{infotypes}{A list, where each component contains information
types for subsequent bootstrap procedures. See 'Details'.}
}
\value{
A list of class \code{"admbsecr.boot"}. Components contain
information such as estimated parameters and standard errors. The
best way to access such information, however, is through the
variety of helper functions provided by the admbsecr package. S3
methods \link{stdEr.admbsecr.boot} and \link{vcov.admbsecr.boot}
can be used to return standard errors and the variance-covariance
matrix of estimated parameters based on the bootstrap procedure.
}
\description{
Carries out a parametric bootstrap, based on a model fitted using
\link{admbsecr}.
}
\details{
For each bootstrap resample, a new population of individuals is
simulated within the mask area. Detections of these individuals are
simulated using the estimated detection function. For detected
individuals, additional informatin is simulated from the estimated
distribution of measurement error. The original model is then
re-fitted to these simulated data, and parameter estimates for each
iteration saved in the component \code{boot} of the returned list.

Note that generic functions \link{stdEr} and \link{vcov} with an
object of class \code{admbsecr.boot} as the main argument will
return standard errors and the variance-covariance matrix for
estimated parameters \emph{based on the bootstrap procedure} (via
the \link{stdEr.admbsecr.boot} and \link{vcov.admbsecr.boot}
methods). For standard errors and the variance-covariance matrix
based on maximum likelihood asymptotic theory, the methods
\link{stdEr.admbsecr} and \link{vcov.admbsecr} must be called
directly.

If \code{infotypes} is provided it should take the form of a list,
where each component is a subset of information types (i.e.,
\code{fit$infotypes}) used to fit the original model. A \code{NULL}
component is associated with no additional information. The
bootstrap procedure is then repeated for each component, only
utilising the appropriate additional information. In practice this
is only useful if the user is looking to investigate the benefits
of including particular information types. The results from these
extra bootstrap procedures can be found in the
\code{boot$extra.boots} component of the returned object.

the bootstrap procedure will be
repeated for each subset; usually this is only useful to
investigate the impact on the density estimator associated with
various combinations of information types. A \code{NULL} component
will bootstrap without any additional information types.
}
\section{Bootstrapping for acoustic surveys}{


For fits based on acoustic surveys where the argument
\code{call.freqs} is provided to the \code{admbsecr} function, the
simulated data allocates multiple calls to the same location based
on an estimated distribution of the call frequencies. Using a
parametric bootstrap is currently the only way parameter
uncertainty can be estimated for such models (see Stevenson et al.,
in prep., for details).
}

\section{Monte Carlo error}{


There will be some error in esimates based on the parametric
bootstrap (e.g., standard errors and estimates of bias) because the
number of bootstrap simulations is not infinite. By default, this
function calculates Monte Carlo error using a bootstrap of the
estimates obtained from the initial bootstrap procedure; see
Equation (9) in Koehler, Brown and Haneuse (2009). Note that this
secondary bootstrap does not require the fitting of any further
models, and so the increased processing time due to this procedure
is negligible.

Monte Carlo error for standard errors and estimates of bias can be
extracted using the function \link{get.mce}.
}
\examples{
\dontrun{
## In practice, N should be >> 100, but this leads to long computation time for a simple example.
boot.fit <- boot.admbsecr(fit = simple.hn.fit, N = 100)
}
}
\references{
Koehler, E., Brown, E., and Haneuse, S. J.-P. A. (2009)
On the assessment of Monte Carlo error in sumulation-based
statistical analyses. \emph{The American Statistician},
\strong{63}: 155--162.

Stevenson, B. C., Borchers, D. L., Altwegg, R., Measey,
G. J., Swift, R. J., and Gillespie, D. M. (in prep.) A general
framework for animal density estimation from acoustic detection
data.
}

