% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot.ascr}
\alias{boot.ascr}
\title{Title}
\usage{
boot.ascr(fit, N, n.cores = 1, M = 10000, infotypes = NULL, seed = NULL)
}
\arguments{
\item{fit}{}

\item{N}{}

\item{n.cores}{}

\item{M}{}

\item{infotypes}{}

\item{seed}{}
}
\value{

}
\description{
Title
}
