% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot.ascr}
\alias{boot.ascr}
\title{Title}
\usage{
boot.ascr(fit, N = 30, n.cores = 1, infotypes = NULL, seed = NULL)
}
\arguments{
\item{fit}{}

\item{N}{}

\item{n.cores}{}

\item{infotypes}{a character vector contains the infotypes will be simulated in the additional bootstraps;
or a list with each element as a character vector, used for doing multiple additional
bootstraps with different settings of infotypes.}

\item{seed}{}
}
\description{
Title
}
