% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.ascr_tmb}
\alias{coef.ascr_tmb}
\title{Extract coefficients from the output of ascr_tmb model}
\usage{
\method{coef}{ascr_tmb}(object, types = NULL, pars = NULL, new_covariates = NULL, ...)
}
\arguments{
\item{object}{a fitted model from fit.ascr.tmb}

\item{types}{a character vector, accept any subset from <'all', 'fitted', 'linked', 'derived'>, default is 'linked'}

\item{pars}{a character vector containing any parameter names}

\item{new_covariates}{a data frame containing the values of covariates of any extended parameter}
}
\value{
a named numeric vector
}
\description{
Extract coefficients from the output of ascr_tmb model
}
