% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.ascr_tmb}
\alias{confint.ascr_tmb}
\title{Extract confidence interval for ascr.tmb models}
\usage{
\method{confint}{ascr_tmb}(
  object,
  types = NULL,
  level = 0.95,
  pars = NULL,
  new_covariates = NULL,
  ...
)
}
\arguments{
\item{object}{a fitted model from fit.ascr.tmb}

\item{types}{a character vector, accept any subset from <'all', 'fitted', 'linked', 'derived'>, default is 'linked'}

\item{level}{confident level, default is 0.95}

\item{pars}{a character vector containing any parameter names}

\item{new_covariates}{a data frame containing the values of covariates of any extended parameter}

\item{...}{}
}
\value{
a matrix
}
\description{
Extract confidence interval for ascr.tmb models
}
