% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert.capt}
\alias{convert.capt}
\alias{convert.capt.to.ascr}
\alias{convert.capt.to.admbsecr}
\alias{convert.capt.to.secr}
\title{Capture history conversion.}
\usage{
convert.capt.to.ascr(capt)

convert.capt.to.admbsecr(capt)

convert.capt.to.secr(capt, traps, capthist = TRUE, cutoff = NULL)
}
\arguments{
\item{capt}{A \code{secr} capture history object for
\code{convert.capt.to.ascr}, or an \code{ascr} capture
history object for \code{convert.capt.to.secr}.}

\item{traps}{}

\item{capthist}{Logical, if \code{TRUE}, a \code{capthist} object
is returned. Otherwise a data frame is returned, which is
suitable for the \code{captures} argument to the
\link{make.capthist} function.}

\item{cutoff}{The signal strength threshold for detection, if
required.}
}
\value{
A capture history object appropriate for analysis using
    either the \code{ascr} or the \code{secr} package.
}
\description{
These functions convert a capture history object between the
structures required for the \code{ascr} and \code{secr}
packages.
}
\examples{
capt <- convert.capt.to.secr(capt = example.data$capt, traps = example.data$traps, cutoff = example.data$cutoff)

}
