% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{convert.mask}
\alias{convert.mask}
\title{Convert mask object}
\usage{
convert.mask(mask)
}
\arguments{
\item{mask}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a mask point. The function
\link{create.mask} will return a suitable
object. Alternatively, this can be a list of such matrices if
detections from multiple detector arrays are being used to fit
a single model.}
}
\value{
An object of class \code{mask}.
}
\description{
Converts an \code{ascr} mask matrix to a \code{secr} mask
object.
}
\details{
The returned object is suitable for use as the \code{mask}
argument of the function \link{secr.fit}.
}
\examples{
mask <- convert.mask(mask = example$mask)

}
