% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert.mask}
\alias{convert.mask}
\title{Convert mask object}
\usage{
convert.mask(mask)
}
\arguments{
\item{mask}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a mask point. The function
\link{create.mask} will return a suitable object.}
}
\value{
An object of class \code{mask}.
}
\description{
Converts an \code{admbsecr} mask matrix to a \code{secr} mask
object.
}
\details{
The returned object is suitable for use as the \code{mask}
argument of the function \link{secr.fit}.
}
\examples{
mask <- convert.mask(mask = example$mask)
}

