\name{create.capt}
\alias{create.capt}
\title{Creating capture history object.
Creates a capture history object to use with the function \code{\link{admbsecr}}.}
\usage{
create.capt(captures, n.traps = NULL)
}
\arguments{
  \item{captures}{A data frame.}

  \item{n.traps}{The total number of traps.}
}
\description{
Creating capture history object. Creates a capture history
object to use with the function \code{\link{admbsecr}}.
}

