% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{create.capt}
\alias{create.capt}
\title{Creating capture history object.}
\usage{
create.capt(
  captures,
  traps = NULL,
  n.traps = NULL,
  n.sessions = NULL,
  mrds.locs = NULL,
  use.name = FALSE
)
}
\arguments{
\item{captures}{}

\item{traps}{}

\item{n.traps}{}

\item{n.sessions}{}

\item{mrds.locs}{}

\item{use.name}{}
}
\value{

}
\description{
Creating capture history object.
}
