% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{create.capt}
\alias{create.capt}
\title{Title}
\usage{
create.capt(
  captures,
  traps = NULL,
  ind_model = NULL,
  n.sessions = NULL,
  n.traps = NULL,
  mrds.locs = NULL
)
}
\arguments{
\item{captures}{}

\item{traps}{}

\item{ind_model}{}

\item{n.sessions}{}

\item{n.traps}{}

\item{mrds.locs}{a list with length of n.sessions with data frames or matrices as components, each data.frame or matrix
contains two columns record the Cartesian coordinates of each call. If a session has no detection,
keep the corresponding component as NULL. If animal.model, then each component must be a data.frame
with 4 columns: "animal_ID", "ID", "mrds_x" and "mrds_y".}
}
\value{

}
\description{
Title
}
\examples{
n_IDs' will be used later to construct the matrices of output
n_animals_call' only used later to check 'mrds.locs'
}
