\name{create.capt}
\alias{create.capt}
\title{Creating capture history object.}
\usage{
create.capt(captures, n.traps = NULL)
}
\arguments{
  \item{captures}{A data frame.}

  \item{n.traps}{The total number of traps.}
}
\description{
Creates a capture history object to use with the function
\code{\link{admbsecr}}.
}

