% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.capt}
\alias{create.capt}
\title{Creating capture history object.}
\usage{
create.capt(captures, n.traps = NULL)
}
\arguments{
\item{captures}{A data frame of capture records, see 'Details' for
the correct format.}

\item{n.traps}{The total number of traps. If \code{NULL} then the
number of traps is assumed to be the largest value in the
\code{traps} column of the \code{captures} argument.}
}
\description{
Creates a capture history object to use with the function
\code{\link{admbsecr}}.
}
\details{
The \code{captures} argument to this function is intended to be of
a similar format to the \code{captures} argument to
\link{make.capthist} in the \link{secr} package. That is, users can
use the same \code{captures} data frame with \code{create.capt} and
\code{make.capthist}, which generate capture histories for use with
the \code{admbsecr} and \link{secr} packages respectively.

As such, the second and fourth columns should provide the ID of the
detection and the trap number of the trap which made the detection
(where the trap number is the row number of the corresponding trap
in the matrix of trap locations). Note that the first and third
columns provide the 'session' and 'occassion' of the detection for
\link{make.capthist}, but as the admbsecr package does not
presently have the capabilities to deal with multi-session or
multi-occassion data, these columns are ignored by
\code{create.capt}.

Additional optional columns can specify the additional information
collected over the course of the survey:
\itemize{
  \item A column named \code{bearing} containing estimated bearings
        from which the detector detected the individual.
  \item A column named \code{dist} containing the estimated
        distance between the individual detected and the detector.
  \item A column named \code{ss} containing the measured signal
        strengh of the detected acoustic signal (only possible when
        detectors are microphones).
  \item A column named \code{toa} containing the measured time of
        arrival (in seconds) since the start of the survey (or some
        other reference time) of the detected acoustic signal (only
        possible when the detectors are microphones).
  \item A column named \code{mrds} containing the \emph{known} (not
        estimated) distance between the individual detected and the
        detector.
}
}

