% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{create.mask}
\alias{create.mask}
\title{Create mask object}
\usage{
create.mask(traps, buffer, ...)
}
\arguments{
\item{traps}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a trap (or
detector). Alternatively, this can be a list of such matrices
if detections from multiple detector arrays are being used to
fit a single model.}

\item{buffer}{The minimum distance between trap locations and the
edge of the generated mask.}

\item{...}{Arguments to be passed to \link{make.mask}.}
}
\value{
An object of class \code{mask}. The object has two useful attributes: \code{area} which gives the area of each grid cell in hectares and \code{buffer} which gives the value of \code{buffer} supplied to this function.
}
\description{
Creates a mask object to use with the function
\code{\link{fit.ascr}}.
}
\examples{
mask <- create.mask(traps = example$traps, buffer = 20)

# calculate the area of the mask (in hectares)
mask_area <- attr(mask, "area") * nrow(mask)

}
\seealso{
\link{convert.mask} to convert a mask compatible with the
    \link{secr} package.
}
