% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detfns.r
\name{detfn}
\alias{detfn}
\title{Calculating the detection function from a fitted model object.}
\usage{
detfn(fit, d)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{d}{A vector of distances.}
}
\value{
A vector of detection probabilities.
}
\description{
Calculates detection probabilities from a fitted model's estimated
detection function.
}
