\name{distances}
\alias{distances}
\title{Calculating distances between mask points and traps for SECR models}
\usage{
  distances(traps, mask)
}
\arguments{
  \item{traps}{matrix containing trap coordinates.}

  \item{mask}{matrix containing mask point coordinates.}
}
\value{
  A matrix.
}
\description{
  Calculates pairwise distances between traps and mask
  points.
}

