% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.capt}
\alias{get.capt}
\title{Extracting capture history object}
\usage{
get.capt(fit)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}
}
\value{
A capture history object.
}
\description{
Extracts the capture history object used in an ascr fit.
}

