% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.capt}
\alias{get.capt}
\title{Extracting the capture histories}
\usage{
get.capt(fit, session = NULL)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{session}{The session from which to extract the capture
histories.}
}
\value{
A capture histories object.
}
\description{
Extracts the capture histories from an ascr fit.
}
