% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.mask}
\alias{get.mask}
\title{Extracting mask point locations}
\usage{
get.mask(fit)
}
\arguments{
\item{fit}{A fitted model from \link{admbsecr}.}
}
\value{
A mask object.
}
\description{
Extracts the mask used in an admbsecr fit.
}

