% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.mask}
\alias{get.mask}
\title{Extracting mask point locations}
\usage{
get.mask(fit, session = NULL)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{session}{The session from which to extract the mask.}
}
\value{
A mask object.
}
\description{
Extracts the mask used in an ascr fit.
}
