% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.traps}
\alias{get.traps}
\title{Extracting trap locations}
\usage{
get.traps(fit)
}
\arguments{
\item{fit}{A fitted model from \link{admbsecr}.}
}
\value{
A traps object.
}
\description{
Extracts the trap locations used in an admbsecr fit.
}

