% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{get.traps}
\alias{get.traps}
\title{Extracting trap locations}
\usage{
get.traps(fit, session = NULL)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{session}{The session from which to extract the trap
locations.}
}
\value{
A traps object.
}
\description{
Extracts the trap locations used in an ascr fit.
}
