% Generated by roxygen2 (4.0.0): do not edit by hand
\name{locations}
\alias{locations}
\title{Plotting estimated locations}
\usage{
locations(fit, id, infotypes = NULL, xlim = range(mask[, 1]),
  ylim = range(mask[, 2]), mask = get.mask(fit), levels = NULL,
  nlevels = 10, density = FALSE, cols = list(combined = "black", capt =
  "purple", bearing = "green", dist = "brown", toa = "blue"), lty = 1,
  show.labels = TRUE, plot.arrows = "bearing" \%in\% fit$infotypes,
  plot.circles = "dist" \%in\% fit$infotypes, arrow.length = NULL,
  show.legend = !add, show.axes = TRUE, add = FALSE)
}
\arguments{
\item{fit}{A fitted model from \link{admbsecr}.}

\item{id}{A numeric vector with row numbers from \code{fit$args$capt},
indicating which individuals' locations are to be plotted.}

\item{infotypes}{A character vector indicating the type(s) of
information to be used when plotting the estimated density of
location.  Elements can be a subset of \code{"capt"},
\code{"bearing"}, \code{"dist"}, \code{"ss"}, \code{"toa"},
\code{"combined"}, and \code{"all"}, where \code{"capt"} shows
estimated location only using detection locations,
\code{"combined"} combines all information types together, and
\code{"all"} plots all possible contour types. When signal strength
information is used in the model fit, \code{"capt"} and \code{"ss"}
are equivalent as the signal strength information is built into the
detection function. By default, only the most informative contour
is plotted, i.e., \code{"capt"} if the model was fitted with no
additional information, and \code{"combined"} otherwise.}

\item{xlim}{A numeric vector of length 2, giving the x coordinate range.}

\item{ylim}{A numeric vector of length 2, giving the y coordinate range.}

\item{mask}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a mask point. The function
\link[admbsecr]{create.mask} will return a suitable object. The
mask used to fit the model \code{fit} will be used by default; this
argument is usually used when estimated location contours need to
be plotted to a higher resolution than this.}

\item{levels}{A numeric vector giving the values to be associated
with the plotted contours.}

\item{nlevels}{The number of contour levels desired. Ignored if
\code{levels} is provided.}

\item{density}{Logical, if \code{TRUE}, the labels on contours (and
the levels specified by \code{levels}) refer to the density of the
estimated distribution of the individual's location. If
\code{FALSE}, the labels on contours (and the levels specified by
\code{levels}) refer to the probability of the individual being
located within the associated contour under the estimated
distribution of the individual's location.}

\item{cols}{A list with named components corresponding to each
contour type (i.e., a subset of \code{"capt"}, \code{"bearing"},
\code{"dist"}, \code{"toa"}, and \code{"combined"}). Each component
provides the colour of the associated contour type (e.g., using a
character string such as \code{"red"}, or a call to the function
\link[grDevices]{rgb}). By default, if only one contour is to be
plotted, it will be plotted in black.}

\item{lty}{The line type of the contours; see \link{par}.}

\item{show.labels}{Logical, if \code{TRUE}, contours are labelled
with the appropriate probability density (if \code{density} is
\code{TRUE}), or the corresponding probability of the individual
being within the associated contour, under the estimated density
(if \code{density} is \code{FALSE}).}

\item{plot.arrows}{Logical, if \code{TRUE}, arrows indicating the
estimated bearing to the individual are plotted from detectors at
which detections were made.}

\item{plot.circles}{Logical, if \code{TRUE}, circles indicating the
estimated distance to the individual are plotted around detectors
at which detections were made.}

\item{arrow.length}{Numeric, providing the length of the arrows
(only used if \code{plot.arrows} is \code{TRUE}).}

\item{show.legend}{Logical, if \code{TRUE}, a legend will be added to
the plot.}

\item{show.axes}{Logical, if \code{TRUE}, axes will be added to the
plot.}

\item{add}{Logical, if \code{TRUE}, contours will be added to an
existing plot.}
}
\description{
Plots estimated densities of animal locations, which are latent
variables in SECR models.
}
\examples{
locations(simple.hn.fit, 1)
locations(simple.hn.fit, 1, levels = c(0.50, 0.90, 0.95))
\dontrun{
fine.mask <- create.mask(example.traps, 20, spacing = 0.2)
locations(bearing.hn.fit, 1, infotypes = "all", mask = fine.mask)
}
}

