% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.acoustic.captures}
\alias{make.acoustic.captures}
\title{Assigning ID numbers to sounds.}
\usage{
make.acoustic.captures(mics, dets, sound.speed)
}
\arguments{
\item{mics}{a matrix containing the coordinates of trap locations.}

\item{dets}{a data frame containing (at least): (i) \code{$toa},
the precise time of arrival of the received sound, and (ii)
\code{$trap} the trap at which the sound was recorded.}

\item{sound.speed}{the speed of sound in metres per second.}
}
\value{
A data frame. Specifically, the \code{dets} dataframe,
now with a new variable, \code{ID}.
}
\description{
Identifies recaptures and assigns ID numbers to sounds recorded for
an SECR model.
}
\details{
Detected sounds are assumed to come from the same animal if times
of arrival at different microphones are closer together than the
time it would take for sound to travel between these microphones.
}
\author{
David Borchers
}

