% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mask.test}
\alias{mask.test}
\title{Testing the mask object for a first calls model.}
\usage{
mask.test(mask, traps, pars, cutoff, lower.cutoff)
}
\arguments{
\item{mask}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a mask point. The function
\link{create.mask} will return a suitable object.}

\item{traps}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a trap (or detector).}

\item{pars}{A named list. Component names are parameter names, and
components are values of parameters at which the mask is to be
tested. Parameters must include \code{b0.ss}, \code{b1.ss}, and
\code{sigma.ss}; see \link{admbsecr} for further details.}

\item{cutoff}{The upper cutoff value; see \code{admbsecr} for
further details.}

\item{lower.cutoff}{The lower cutoff value; see \code{admbsecr} for
further details.}
}
\description{
Creates a diagnostic plot that can be used to test the adequacy of
a mask for a first calls model.
}

