% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mask.test}
\alias{mask.test}
\title{Testing the mask object for a first calls model.}
\usage{
mask.test(mask, traps, pars, cutoff, lower.cutoff)
}
\arguments{
\item{mask}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a mask point. The function
\link{create.mask} will return a suitable object.}

\item{traps}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a trap (or detector).}
}
\description{
Creates a diagnostic plot that can be used to test the adequacy of
a plot for a first calls model.
}

