% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_in_fit.R
\name{outFUN}
\alias{outFUN}
\title{D' is only session not trap extendable, we could take the first observation in each session}
\usage{
outFUN(
  data.par,
  data.full,
  data.traps,
  data.mask,
  data.dists.thetas,
  detfn,
  param.og,
  param.og.4cpp,
  o,
  tmb_output_og,
  opt,
  name.fixed.par,
  name.extend.par,
  dims,
  DX.full,
  DX.mask,
  fix.input,
  bucket_info,
  cue.rates,
  mean.cue.rates,
  A,
  survey.length,
  sound.speed,
  par.extend,
  arg.input,
  fgam,
  gam_output,
  is.scale,
  ss.link,
  cutoff
)
}
\description{
D' is only session not trap extendable, we could take the first observation in each session
}
