% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_in_fit.R
\name{par.extend.fun}
\alias{par.extend.fun}
\title{fgam' is an object for compatibility with ADMB model
animal_ID',}
\usage{
par.extend.fun(
  par.extend,
  data.full,
  data.mask,
  animal.model,
  dims,
  fulllist.par,
  extra_args
)
}
\description{
fgam' is an object for compatibility with ADMB model
animal_ID',
}
