\name{se.correct}
\alias{se.correct}
\title{SECR standard error correction}
\usage{
  se.correct(fit, calls, size)
}
\arguments{
  \item{fit}{an \code{admbsecr} model fit.}

  \item{calls}{a vector containing call frequencies from
  monitored individuals.}

  \item{size}{number of bootstrap samples.}
}
\description{
  Corrects standard errors from an \code{admbsecr} fit.
}

