\name{se.correct}
\alias{se.correct}
\title{SECR standard error correction}
\usage{
  se.correct(fit, rates, size)
}
\arguments{
  \item{fit}{an \code{admbsecr} model fit.}

  \item{rates}{a vector containing call frequencies from
  monitored individuals.}
}
\description{
  Corrects standard errors from an \code{admbsecr} fit.
}

