\name{show.detfn}
\alias{show.detfn}
\title{Plotting a detection function.}
\usage{
show.detfn(fit, xlim = NULL, ylim = c(0, 1), main = NULL,
  xlab = "Distance (m)", ylab = "Detection probability", add = FALSE, ...)
}
\arguments{
  \item{fit}{A fitted model from
  \link[admbsecr]{admbsecr}.}

  \item{xlim}{A vector with two elements, giving the x-axis
  limits. If \code{NULL}, this is set to the buffer of the
  mask.}

  \item{ylim}{A vector with two elements, giving the y-axis
  limits.}

  \item{main}{The main title.}

  \item{xlab}{The x-axis label.}

  \item{ylab}{The y-axis label.}

  \item{add}{Logical, if \code{TRUE}, the estimated
  detection function is added to an existing plot.}

  \item{...}{Further arguments to be passed to
  \link[graphics]{lines}.}
}
\description{
Plots an estimated detection function from a model fit
generated by \link[admbsecr]{admbsecr}.
}

