% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detfns.r
\name{show.detfn}
\alias{show.detfn}
\title{Plotting a detection function.}
\usage{
show.detfn(fit, xlim = NULL, ylim = c(0, 1), main = NULL,
  xlab = "Distance (m)", ylab = "Detection probability", add = FALSE,
  ...)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{xlim}{A vector with two elements, giving the x-axis
limits. If \code{NULL}, this is set to the buffer of the mask.}

\item{ylim}{A vector with two elements, giving the y-axis limits.}

\item{main}{The main title (on top) using font, size (character
    expansion) and color \code{par(c("font.main", "cex.main", "col.main"))}.}

\item{xlab}{X axis label using font, size and color 
    \code{par(c("font.lab", "cex.lab", "col.lab"))}.}

\item{ylab}{Y axis label, same font attributes as \code{xlab}.}

\item{add}{Logical, if \code{TRUE}, the estimated detection
function is added to an existing plot.}

\item{...}{Further arguments to be passed to \link{lines}.}
}
\description{
Plots an estimated detection function from a model fit generated by
\link{fit.ascr}.
}
\examples{
## Comparison of two detection functions fitted to the same data
show.detfn(example$fits$simple.hn, main = "Detection function comparison")
show.detfn(example$fits$simple.hr, add = TRUE, col = "blue")
legend("topright", legend = c("Half normal", "Hazard rate"), lty = 1, col = c("black", "blue"), bg = "white")

}
