% Generated by roxygen2 (4.0.0): do not edit by hand
\name{show.detfn}
\alias{show.detfn}
\title{Plotting a detection function.}
\usage{
show.detfn(fit, xlim = NULL, ylim = c(0, 1), main = NULL,
  xlab = "Distance (m)", ylab = "Detection probability", add = FALSE, ...)
}
\arguments{
\item{xlim}{A vector with two elements, giving the x-axis
limits. If \code{NULL}, this is set to the buffer of the mask.}

\item{ylim}{A vector with two elements, giving the y-axis limits.}

\item{add}{Logical, if \code{TRUE}, the estimated detection
function is added to an existing plot.}

\item{...}{Further arguments to be passed to \link{lines}.}

\item{fit}{A fitted model from \link{admbsecr}.}

\item{main}{The main title (on top) using font and size (character
    expansion) \code{par("font.main")} and color
    \code{par("col.main")}.}

\item{xlab}{X axis label using font and character expansion
    \code{par("font.lab")} and color \code{par("col.lab")}.}

\item{ylab}{Y axis label, same font attributes as \code{xlab}.}
}
\description{
Plots an estimated detection function from a model fit generated by
\link{admbsecr}.
}
\examples{
## Comparison of two detection functions fitted to the same data
show.detfn(simple.hn.fit, main = "Detection function comparison")
show.detfn(simple.hr.fit, add = TRUE, col = "blue")
legend("topright", legend = c("Half normal", "Hazard rate"), lty = 1, col = c("black", "blue"), bg = "white")
}

