% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surveyplots.r
\name{show.detsurf}
\alias{show.detsurf}
\title{Plotting the detection probability surface}
\usage{
show.detsurf(fit, ...)
}
\arguments{
\item{fit}{A fitted model from \link{admbsecr}.}

\item{...}{Arguments to be passed to \link{plot}.}
}
\description{
Plots the detection probability surface, based on trap locations
and the estimated detection function from a model fitted using
\link{admbsecr}.
}
\examples{
show.detsurf(example$fits$simple.hn)
}

