% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.r
\name{show.Dsurf}
\alias{show.Dsurf}
\title{Plotting an estimated density surface.}
\usage{
show.Dsurf(
  fit,
  session = 1,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  plot.contours = TRUE
)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{session}{The session with the detector array and invidual(s)
to be plotted (for multi-session models only).}

\item{xlim}{A numeric vector of length 2, giving the x coordinate
range.}

\item{ylim}{A numeric vector of length 2, giving the y coordinate
range.}

\item{zlim}{A numberic vector of length 2, giving the range of
density contours.}

\item{plot.contours}{Logical, if \code{TRUE}, contours are
plotted.}
}
\description{
Plots density surface estimated by a model fitted with the function
\link{fit.ascr}.
}
\examples{
simple.capt <- example.data$capt[1]
cov.df <- data.frame(x = example.data$mask[, 1]/1000, y = example.data$mask[, 2]/1000)
fit <- fit.ascr(capt = simple.capt, traps = example.data$traps, mask = example.data$mask,
                fix = list(g0 = 1), ihd.opts = list(model = ~ x + y,
                                                    covariates = cov.df))
show.Dsurf(fit)
}
