% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.r
\name{show.Dsurf}
\alias{show.Dsurf}
\title{Plotting an estimated density surface.}
\usage{
show.Dsurf(
  fit,
  session = 1,
  newdata = NULL,
  show.cv = FALSE,
  unsuitable = NULL,
  xlim = NULL,
  ylim = NULL,
  zlim = NULL,
  scale = 1,
  plot.contours = TRUE,
  add = FALSE
)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{session}{The session with the detector array and invidual(s)
to be plotted. Ignored if the \code{newdata} argument is
provided.}

\item{newdata}{A data frame including new mask points and covariate
values, from which to estimate and plot density estimates
for. This allows, for example, estimates to be provided for new
regions not included in the mask used to fit the model. Two
columns, named \code{x} and \code{y}, must be included,
providing the x- and y-coordinates of the new mask
points. Additional columns must provide the covariates used to
fit the model.}

\item{show.cv}{Logical. If true, the CV of the density estimate is
plotted rather than the estimate itself. At present, this will
only work if \code{newdata} is also provided.}

\item{unsuitable}{A data frame with two columns, named \code{x} and
\code{y}, providing mask points in unsuitable habitat. These
mask points are allocated a density of zero in the plot.}

\item{xlim}{A numeric vector of length 2, giving the x coordinate
range.}

\item{ylim}{A numeric vector of length 2, giving the y coordinate
range.}

\item{zlim}{A numberic vector of length 2, giving the range of
density contours.}

\item{scale}{By default, density is in animals per hectare. The
plotted values are multiplied by this argument, allowing for
user-specified units. For example, setting \code{scale = 100}
results in densities plotted as animals per square kilometre.}

\item{plot.contours}{Logical, if \code{TRUE}, contours are plotted.}

\item{add}{Logical, if \code{TRUE}, contours will be added to an
existing plot.}
}
\description{
Plots density surface estimated by a model fitted with the function
\link{fit.ascr}.
}
\examples{
simple.capt <- example.data$capt[1]
cov.df <- data.frame(x = example.data$mask[, 1]/1000, y = example.data$mask[, 2]/1000)
fit <- fit.ascr(capt = simple.capt, traps = example.data$traps, mask = example.data$mask,
                fix = list(g0 = 1), ihd.opts = list(model = ~ x + y,
                                                    covariates = cov.df))
show.Dsurf(fit)
}
