% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyplots.r
\name{show.survey}
\alias{show.survey}
\title{Plotting mask and trap layout}
\usage{
show.survey(fit, session = 1, ...)
}
\arguments{
\item{fit}{A fitted model from \link{fit.ascr}.}

\item{session}{The session for which the mask point and trap
locations are to be plotted.}

\item{...}{Further arguments to be passed to \link{plot}.}
}
\description{
Plots the mask points and trap locations used in a model fitted
with the function \link{fit.ascr}.
}
\examples{
show.survey(example$fits$simple.hn)

}
