% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcapt.r
\name{sim.capt}
\alias{sim.capt}
\title{Simulating SECR data}
\usage{
sim.capt(fit = NULL, traps = NULL, mask = NULL, popn = NULL,
  infotypes = character(0), detfn = "hn", pars = NULL,
  ss.opts = NULL, cue.rates = NULL, survey.length = NULL,
  freq.dist = "edf", sound.speed = 330, ihd.surf = NULL,
  test.detfn = FALSE, first.only = FALSE, keep.locs = FALSE,
  keep.ids = FALSE, ...)
}
\arguments{
\item{fit}{A fitted \code{ascr} model object which provides the
additional information types, detection function, and parameter
values from which to generate capture histories.}

\item{traps}{A matrix with two columns. Each row provides Cartesian
coordinates for the location of a trap (or
detector). Alternatively, this can be a list of such matrices
if detections from multiple detector arrays are being used to
fit a single model.}

\item{mask}{A matrix with two columns, providing x- and
y-coordinates respectively. The extreme x- and y-coordinates
define the rectangle in which individuals' locations are
simulated.}

\item{popn}{A matrix with two columns, providing x- and
y-coordinates of the animal locations.}

\item{infotypes}{A character vector indicating the type(s) of
additional information to be simulated. Elements can be a
subset of \code{"bearing"}, \code{"dist"}, \code{"toa"}, and
\code{"mrds"} (NOTE: \code{"mrds"} not yet implemented). If
signal strength information is required, specify \code{detfn}
as \code{"ss"} rather than including it here.}

\item{detfn}{A character string specifying the detection function
to be used. Options are "hn" (halfnormal), "hr" (hazard rate),
"th" (threshold), "lth" (log-link threshold), or "ss" (signal
strength).}

\item{pars}{A named list. Component names are parameter names, and
each component is the value of the associated parameter. A
value for the parameter \code{D}, animal density (or call
density, if it an acoustic survey) must always be provided,
along with values for parameters associated with the chosen
detection function and additional information type(s).}

\item{ss.opts}{Options for models using the signal strength
detection function. See 'Details' below.}

\item{cue.rates}{A vector of call frequencies from which a
distribution for the number of emitted calls for each
individual is fitted. If scalar, all individuals make the same
number of calls. If \code{Inf}, \code{first.only} must be
\code{TRUE}, and all individuals keep making calls until the
first is detected.}

\item{survey.length}{The length of a cue-based survey.}

\item{freq.dist}{A character string, either \code{"edf"} or
\code{"norm"}, which specifies how the distribution function of
the call frequencies should be estimated. If \code{"edf"}, then
the distribution of call frequencies is estimated using the
empirical distribution function. If \code{"norm"}, then a
normal distribution is fitted to the call frequencies using the
sample mean and variance. If \code{cue.rates} is scalar then
this is ignored, and all individuals make the same number of
calls. See 'Details' below for information on how call
frequencies are rounded.}

\item{sound.speed}{The speed of sound in metres per second,
defaults to 330 (the speed of sound in air). Only used when
\code{"toa"} is a component name of \code{capt}.}

\item{ihd.surf}{A list with a vector for each session containing a
value for animal density at each mask point. Only required to
simulate inhomogeneous density surfaces.}

\item{test.detfn}{Logical value, if \code{TRUE}, tests detection
function to aid debugging.}

\item{first.only}{Only keep the first detection for each
individual.}

\item{keep.locs}{Logical, if \code{TRUE}, the locations of
individuals in the simulated population are retained. In this
case, the capture histories and auxiliary information are kept
in a component \code{capt} of the returned list, locations of
detected individuals are kept in a component \code{capt.locs},
and locations of all individuals in the population are kept in
a component \code{popn.locs}. In this case, the capture
histories and auxiliary information are kept in a component
\code{capt} of the returned list, and ID numbers of detected
individuals are kept in a component \code{capt.ids}}

\item{keep.ids}{Logical, if \code{TRUE}, the ID number of detected
individuals are retained.}

\item{...}{Other arguments (mostly for back-compatibility).}
}
\value{
A list with named components, each corresponding to a data
    type listed in \code{infotypes}. Each component is a matrix
    where each row corresponds to each detected individual, and
    each column corresponds to a trap (or detector). The elements
    in the matrix indicate detection, and provide simulated values
    of the additional information requested. This object can be
    used as the \code{capt} argument for the function
    \link{fit.ascr}.
}
\description{
Simulates SECR capture histories and associated additional
information in the correct format for use with the function
\link{fit.ascr}. If \code{fit} is provided then no other arguments
are required. Otherwise, at least \code{traps}, \code{mask}, and
\code{pars} are needed.
}
\details{
See documentation for the function \link{fit.ascr} for information
on the parameters corresponding to the different detection
functions, and to different types of additional information.

Simulated call frequencies are not always integers, e.g. when
\code{freq.dist} is \code{"norm"}, or when \code{freq.dist} is
\code{"edf"} and the call frequencies used to fit the model
\code{fit} are not all integers. In this case, if \code{freq.dist}
is \code{"edf"}, then simulated call frequencies are rounded at
random as follows: Let \eqn{x} be the fraction part of the number,
then the call frequency is rounded up with probability \eqn{x} and
rounded down with probability \eqn{1 - x}. For example, a value of
8.1 will be rounded to 9 with probability 0.1, and rounded to 8
with probability 0.9.

If \code{cue.rates} is \code{Inf} then all simulated individuals
will be detected. To generate sensible capture histories then both
a lower and upper cutoff must be supplied in \code{ss.opts}. In
this case, individuals continue to emit calls until one is detected
above the lower cutoff by at least one microphone. This individual
is then incorporated into the resulting capture history only if the
loudest received signal strength is also above the upper cutoff.
}
\examples{
## Simulating based on model fits.
simple.capt <- sim.capt(example$fits$simple.hn)
bearing.capt <- sim.capt(example$fits$bearing.hn)
## Simulating from provided parameter values.
new.capt <- sim.capt(traps = example$traps, mask = example$mask, infotypes = c("bearing", "dist"), detfn = "hr",
                     pars = list(D = 2500, g0 = 0.9, sigma = 3, z = 2, kappa = 50, alpha = 10))

}
