% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{simple.hr.fit}
\alias{simple.hr.fit}
\title{An example model object}
\format{A list of class \code{"admbsecr"}.}
\usage{
simple.hr.fit
}
\description{
This is the model object that results when the
\link{admbsecr} function is run with
\link{example.capt}\code{["bincapt"]},
\link{example.traps}, \link{example.mask}, and
\code{"hr"} set as the arguments \code{capt}, \code{traps},
\code{mask}, and \code{detfn}, respectively.
}
\keyword{datasets}

