% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tmb.R
\name{stdEr.ascr_tmb}
\alias{stdEr.ascr_tmb}
\title{Extract standard errors of the estimated parameters from ascr.tmb models}
\usage{
\method{stdEr}{ascr_tmb}(object, types = NULL, pars = NULL, new_covariates = NULL, ...)
}
\arguments{
\item{object}{a fitted model from fit.ascr.tmb}

\item{types}{a character vector, accept any subset from <'all', 'fitted', 'linked', 'derived'>, default is 'linked'}

\item{pars}{a character vector containing any parameter names}

\item{new_covariates}{a data frame containing the values of covariates of any extended parameter}
}
\value{
a named numeric vector
}
\description{
Extract standard errors of the estimated parameters from ascr.tmb models
}
