% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.ascr_tmb}
\alias{vcov.ascr_tmb}
\title{Extract variance covariance matrix of the estimated parameters from ascr models}
\usage{
\method{vcov}{ascr_tmb}(
  object,
  types = NULL,
  pars = NULL,
  new_covariates = NULL,
  show_fixed_par = TRUE,
  ...
)
}
\arguments{
\item{object}{a fitted model from fit.ascr().}

\item{types}{a character vector, the same as "coef.ascr_tmb()".}

\item{pars}{a character vector, the same as "coef.ascr_tmb()".}

\item{new_covariates}{a data frame, the same as "coef.ascr_tmb()".}

\item{show_fixed_par}{a logical value. To control whether to include the fixed parameters in the covariance matrix.
It is TRUE by default.}

\item{...}{}
}
\value{
a list with matrices as its elements if multiple 'types'. a matrix if only one 'types'.
}
\description{
Extract variance covariance matrix of the estimated parameters from ascr models
}
