% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.ascr_tmb}
\alias{vcov.ascr_tmb}
\title{Extract variance covariance matrix of the estimated parameters from ascr.tmb models}
\usage{
\method{vcov}{ascr_tmb}(object, types = NULL, pars = NULL, new_covariates = NULL, ...)
}
\arguments{
\item{object}{a fitted model from fit.ascr.tmb}

\item{types}{a character vector, accept any subset from <'all', 'fitted', 'linked', 'derived'>, default is 'linked'}

\item{pars}{a character vector containing any parameter names}

\item{new_covariates}{a data frame containing the values of covariates of any extended parameter}
}
\value{
a list with matrices as its elements if multiple 'types'. a matrix if only one 'types'.
}
\description{
Extract variance covariance matrix of the estimated parameters from ascr.tmb models
}
